vi keymap
The layers have dvorak on layer 0; number, symbols, and navigation on
layer 1; and function keys and other navigation on layer 2. Temporarily,
the layout for stenography is on layer 3. If more keys are desired, this
will be moved to layer 5, and then the new layer 3 and layer 4 can be
populated.

The thumb clusters have mod-tap except the two inner-most ones have
layer-tap. From left to right, the tap keys on the thumb cluster are as
follows: Esc, Tab, Return, Space, Backspace, &none. The mods and layers
in order are these: LCTRL, LGUI, momentary layer 2, momentary layer 1,
LALT, and LSHIFT.

There is a cluster of numbers on the left side of layer 1 and symbols on
the top of a standard US qwerty keybaord. Use the shift key to access
the symbols on the shifted numbers. The right side has navigation keys
on the first two rows (and a forward slash all the way to the right) and
parentheses and brackets on the bottom row.

Layer 2 mirrors layer 1 in in that the function keys are in the same
positions but F10 through F12 are placed just to the right, as well.
Print screen is just to the left. There are control keys on the right
side of layer 2 that control volume and bluetooth. There is also Insert
and Delete on the right side.

The two most awkward keys on layer 1 and 2 (top-right and top-left
resp.) may be used in the future for infrequently used keys. Currently,
there is no perceived need for them. In order to activate the
stenography keyboard, use the top-right key on layer 2.