try homerow again, reorg
homerow mods on bottom row so that they are less frequent with drovrak.

reorganize layers. Each layer has mods on the bottom left (same
position as the homerow on the left on BASE layer).

layers are organized into BASE (dvorak), NUM (numbers and symbols), FN
(function and utilities), NAV (navigation including LG(#) to easily
switch to a workspace on a tiling wm), and STN (stenography).